split the Diary service in two parts (cached and not cached); using the non cached version for now
