+++ src/MySelfLog.DiaryCache/MySelfLog.DiaryCache.csproj
@@ -49,6 +49,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DiaryService.cs" />
     <Compile Include="DiaryCacheService.cs" />
     <Compile Include="DiaryEvent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
