+++ src/MySelfLog.Host/App.config
@@ -5,7 +5,7 @@
   </configSections>
   <appSettings>
     <add key="EventStoreUserName" value="admin" />    
-    <add key="EventStorePassword" value="changeit" />
+    <add key="EventStorePassword" value="cicciomariano" />
     <add key="EventStoreLink" value="tcp://admin:cicciomariano@eventstore:1113" />
     <add key="ElasticSearchLink" value="http://elasticsearch:9200" /> 
   </appSettings>
